blueprints/genai-marketing-campaigns/backend-img-indexing/requirements.txt
aws-cdk-lib==2.185.0
cdk-nag==2.28.196
aws_cdk.aws_lambda_python_alpha===2.100.0a0
constructs>=10.0.0,<11.0.0
boto3==1.28.63